Renames the scraper package to ingest to reflect its more general nature.
