+++ ingest/models/__init__.py
@@ -19,4 +19,4 @@
 """The core models specific to the ingest platform, but agnostic across
 scraper vendors and regions."""
 
-from scraper.models import scrape_session
+from ingest.models import scrape_session
