+++ ingest/us_ny/us_ny_snapshot.py
