+++ tests/context.py
@@ -20,6 +20,6 @@ sys.path.insert(1, os.path.join(
     os.path.dirname(os.path.realpath(__file__)), '../lib'))
 
 import calculator
+import ingest
 import models
-import scraper
 import utils
